get rid of "browser" blueprint
