+++ scout/pileup/partial.py
