+++ scout/pileup/templates/pileup.html
@@ -22,7 +22,7 @@
               viz: pileup.viz.genome(),
               isReference: true,
               data: pileup.formats.twoBit({
-                url: '{{ url_for("browser.remote_static", file="/mnt/hds/proj/bioinfo/MIP_ANALYSIS/references/hg19.2bit") }}'
+                url: '{{ url_for("pileup.remote_static", file="/mnt/hds/proj/bioinfo/MIP_ANALYSIS/references/hg19.2bit") }}'
               }),
               name: 'Reference'
             },
@@ -37,7 +37,7 @@
             {
               viz: pileup.viz.genes(),
               data: pileup.formats.bigBed({
-                url: '{{ url_for("browser.remote_static", file="/mnt/hds/proj/bioinfo/MIP_ANALYSIS/references/collapsedEnsemblExons.bb") }}'
+                url: '{{ url_for("pileup.remote_static", file="/mnt/hds/proj/bioinfo/MIP_ANALYSIS/references/collapsedEnsemblExons.bb") }}'
               }),
               name: 'Genes'
             },
@@ -45,8 +45,8 @@
               {
                 viz: pileup.viz.coverage(),
                 data: pileup.formats.bam({
-                  url: '{{ url_for("browser.remote_static", file=alignment.bam) }}',
-                  indexUrl: '{{ url_for("browser.remote_static", file=alignment.bai) }}'
+                  url: '{{ url_for("pileup.remote_static", file=alignment.bam) }}',
+                  indexUrl: '{{ url_for("pileup.remote_static", file=alignment.bai) }}'
                 }),
                 name: 'Coverage'
               },
@@ -55,8 +55,8 @@
                     viewAsPairs: true
                 }),
                 data: pileup.formats.bam({
-                  url: '{{ url_for("browser.remote_static", file=alignment.bam) }}',
-                  indexUrl: '{{ url_for("browser.remote_static", file=alignment.bai) }}'
+                  url: '{{ url_for("pileup.remote_static", file=alignment.bam) }}',
+                  indexUrl: '{{ url_for("pileup.remote_static", file=alignment.bai) }}'
                 }),
                 cssClass: 'normal',
                 name: '{{ alignment.sample }}'
