+++ scout/settings.py
@@ -39,7 +39,6 @@ class BaseConfig(object):
                 ('scout.login.login', None),
                 ('scout.user.user', None),
                 ('scout.api.api', None),
-                ('scout.browser.browser', None),
                 ('scout.pileup.pileup_bp', None),
                 ('chanjo_report.server.blueprints.report_bp', '/reports')]
